+++ index.html
@@ -245,6 +245,6 @@
         <button class="error__button" type="button">Попробовать снова</button>
       </div>
     </template>
-    <script src="./js/main.js"></script>
+    <script src="./js/main.js" type="module"></script>
   </body>
 </html>
